requirements.txt
pyTelegramBotAPI==4.22.0
flask==3.0.3
yt-dlp[default]==2024.10.22
openai==1.12.0
moviepy==1.0.3
instagrapi==2.2.1
redis==5.0.8
rq==1.16.2
python-dotenv==1.0.1
tonsdk==1.0.15
pytonconnect==0.2.1
tiktoken==0.7.0
ffmpeg-python==0.2.0
pydantic==2.11.7
requests==2.31.0
